timer: set AccuracySec to 6h and RandomizedDelaySec to 1h
The previous configuration caused the updatedb timer to run
potentially in the late morning. However this should be avoided to not
affect normal users. Furthermore, the low setting of AccuracySet to
20m caused a low chance of systemd to coalesce the updatedb timer with
other timers.

Increase the AccuracySec to 6h so that systemd is more likely to be
able to coalesce the timer with others in an effort to optimize power
consumption.

Furthermore, lower RandomizedDelaySec to 1h, to ensure that the
plocate updatedb job runs before 07:00 in the morning (and after
midnight).