Add an undocumented flag --ignore-visibility.
This is mostly useful in the case where you are root, but the directories
in question no longer exist (or were never there to begin with) and you
still want to search in the database, or dump all of it using plocate ''.
Obviously, this needs to drop the sgid flag like --debug already does,
so we do that.

Using this together with -e/--existing doesn't make any sense,
but we don't have any explicit checks for it.