Remove ConditionACPower=true.
A lot of people use laptops like mobile phones, they use it mostly being on
battery and plug it over night, when the device is turned off, or when battery
level is really down. Then most likely the laptop is on battery when the
plocate update job is due. If you are on battery, systemd services with
ConditionACPower=true do silently nothing, the job is not started and there is
only a minor log notice about it which will go unnoticed most likely. As
plocate is really fast, remove ConditionACPower=true from
plocate-updatedb.service.in.